Revert "Download manager"
