+++ src/pages/download-manager/download-manager.module.ts
@@ -1,4 +1,3 @@
-import { DownloadsTabPage } from './downloads-tab/downloads-tab';
 import { NoDownloadsComponent } from './no-downloads/no-downloads';
 import { NgModule } from '@angular/core';
 import { IonicPageModule } from 'ionic-angular';
@@ -9,8 +8,7 @@ import { TranslateModule } from '@ngx-translate/core';
 @NgModule({
   declarations: [
     DownloadManagerPage,
-    NoDownloadsComponent,
-    DownloadsTabPage
+    NoDownloadsComponent
   ],
   imports: [
     IonicPageModule.forChild(DownloadManagerPage),
